requirements.txt
annotated-types==0.7.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
et_xmlfile==2.0.0
filelock==3.19.1
fonttools==4.59.2
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
idna==3.10
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.10.6
mpmath==1.3.0
networkx==3.4.2
nibabel==5.3.2
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
openpyxl==3.1.5
packaging==25.0
pandas==2.3.2
pillow==11.3.0
platformdirs==4.4.0
protobuf==6.32.1
pyarrow==21.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
rtdl==0.0.13
scikit-learn==1.7.2
scipy==1.15.3
sentry-sdk==2.37.1
simpleitk==2.5.2
six==1.17.0
smmap==5.0.2
sympy==1.14.0
threadpoolctl==3.6.0
torch==1.13.0
torchvision==0.14.0
tqdm==4.67.1
triton==3.4.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wandb==0.21.4
